Increase sweep_channels timeout to 30 seconds
Prior to this commit, the timeout period to wait for the control board
to complete a `sweep_channels` call was 5 seconds.  While 5 seconds is
long enough for a scan of a single channel, now that the control
board sweeps all channels in a single call, the process may take over 5
seconds.

This commit increases the `sweep_channels` timeout to 30 seconds to
compensate for this.